+++ jsonschema-examples/src/test/java/com/github/victools/jsonschema/examples/ExampleTest.java
@@ -40,6 +40,7 @@ public class ExampleTest {
             InheritanceRefExample.class,
             JacksonDescriptionAsTitleExample.class,
             JacksonSubtypeDefinitionExample.class,
+            NamingStrategyExample.class,
             SingleArrayItemExample.class,
             StrictTypeInfoExample.class,
             SubtypeLookUpExample.class,