+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/impl/SchemaCleanUpUtils.java
@@ -24,6 +24,7 @@ import com.github.victools.jsonschema.generator.SchemaGeneratorConfig;
 import com.github.victools.jsonschema.generator.SchemaKeyword;
 import com.github.victools.jsonschema.generator.SchemaVersion;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -270,8 +271,11 @@ public class SchemaCleanUpUtils {
             return null;
         }
         return () -> {
+            Set<JsonNode> arrayItems = new LinkedHashSet<>();
+            arrayNodesToMerge.forEach(node -> node.forEach(arrayItems::add));
+
             ArrayNode mergedArrayNode = this.config.createArrayNode();
-            arrayNodesToMerge.forEach(node -> node.forEach(mergedArrayNode::add));
+            arrayItems.forEach(mergedArrayNode::add);
             return mergedArrayNode;
         };
     }
@@ -644,21 +648,11 @@ public class SchemaCleanUpUtils {
      */
     private void addTypeInfoWhereMissing(ObjectNode schemaNode, String typeTagName, boolean considerNullType,
             Map<String, SchemaKeyword> reverseTagMap) {
-        if (schemaNode.has(typeTagName)) {
-            // explicit type indication is already present
-            return;
-        }
-        List<String> impliedTypes = reverseTagMap.entrySet().stream()
-                .filter(entry -> schemaNode.has(entry.getKey()))
-                .flatMap(entry -> entry.getValue().getImpliedTypes().stream())
-                .distinct()
-                .sorted()
-                .map(SchemaKeyword.SchemaType::getSchemaKeywordValue)
-                .collect(Collectors.toList());
+        List<String> impliedTypes = this.collectImpliedTypes(schemaNode, typeTagName, reverseTagMap);
         if (impliedTypes.isEmpty()) {
             return;
         }
-        if (considerNullType) {
+        if (considerNullType && !this.config.shouldAlwaysWrapNullSchemaInAnyOf()) {
             impliedTypes.add(SchemaKeyword.SchemaType.NULL.getSchemaKeywordValue());
         }
         if (impliedTypes.size() == 1) {
@@ -666,6 +660,24 @@ public class SchemaCleanUpUtils {
         } else {
             impliedTypes.forEach(schemaNode.putArray(typeTagName)::add);
         }
+        if (considerNullType && this.config.shouldAlwaysWrapNullSchemaInAnyOf()) {
+            // since version 4.37.0
+            SchemaGenerationContextImpl.makeNullable(schemaNode, this.config);
+        }
+    }
+
+    private List<String> collectImpliedTypes(ObjectNode schemaNode, String typeTagName, Map<String, SchemaKeyword> reverseTagMap) {
+        if (schemaNode.has(typeTagName)) {
+            // explicit type indication is already present
+            return Collections.emptyList();
+        }
+        return reverseTagMap.entrySet().stream()
+                .filter(entry -> schemaNode.has(entry.getKey()))
+                .flatMap(entry -> entry.getValue().getImpliedTypes().stream())
+                .distinct()
+                .sorted()
+                .map(SchemaKeyword.SchemaType::getSchemaKeywordValue)
+                .collect(Collectors.toList());
     }
 
     /**